pyproject.toml
[project]
name = "pepper"
version = "0.1.0"
description = "Pepper Second Brain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "apscheduler>=4.0.0a5",
    "discord.py>=2.7.0",
    "filelock>=3.25.2",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "jinja2>=3.1.6",
    "mcp>=1.9.0",
    "psutil>=7.0.0",
    "pyqmd>=0.1.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "sqlalchemy>=2.0.0",
    "typer>=0.24.1",
    "watchdog>=6.0.0",
]

[project.scripts]
pepper = "pepper.cli:app"
pepper-channel = "pepper.channel.server:main"
pepper-discord = "pepper.integrations.discord.mcp_server:run"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pip-audit>=2.6.0",
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "pytest-drill-sergeant>=0.4.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.9",
    "xenon>=0.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pepper"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "B",     # flake8-bugbear
    "Q",     # flake8-quotes
    "SIM",   # flake8-simplify
    "C90",   # mccabe complexity
    "ARG",   # flake8-unused-arguments
    "TID",   # flake8-tidy-imports
    "UP",    # pyupgrade
    "PL",    # pylint
    "PERF",  # perflint
    "D",     # pydocstyle
    "RUF",   # ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
    "D200", "D205", "D400", "D401", "D415",
    "ARG001", "ARG002",
]
# Discord integration uses relative imports by design (sub-package)
"src/pepper/integrations/discord/**/*.py" = ["TID252"]
# Skills contain standalone scripts not part of the main package
"src/pepper/skills/**/*.py" = ["D", "E501", "TID252", "ARG", "PL", "N", "C90"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.mccabe]
max-complexity = 15

# ---------------------------------------------------------------------------
# MyPy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "discord.*",
    "apscheduler.*",
    "mcp.*",
    "pyqmd.*",
    "filelock.*",
    "google.*",
    "googleapiclient.*",
    "google_auth_oauthlib.*",
    "httpx_sse.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (isolated, no I/O)",
    "integration: Integration tests (may use localhost services)",
    "e2e: End-to-end tests (full system, may use network)",
    "slow: tests that invoke real Claude Code sessions (cost API tokens)",
]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-p", "no:drill_sergeant",  # TODO: enable after tests have AAA structure
]

# ---------------------------------------------------------------------------
# Coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "__repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"

# ---------------------------------------------------------------------------
# drill-sergeant
# ---------------------------------------------------------------------------
[tool.drill_sergeant]
enabled = false  # Phase 2: enable after tests are reorganized with markers and AAA structure
enforce_markers = false
enforce_markers_severity = "error"
enforce_file_length = true
enforce_file_length_severity = "warn"
max_file_length = 500
